Refactor: UserIdNickname 프로젝션 위치 이동
