+++ src/main/java/project/flipnote/user/model/UserIdNickname.java
@@ -1,4 +1,4 @@
-package project.flipnote.group.model;
+package project.flipnote.user.model;
 
 public interface UserIdNickname {
 	Long getId();
