+++ src/main/java/project/flipnote/user/repository/UserProfileRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import project.flipnote.group.model.UserIdNickname;
+import project.flipnote.user.model.UserIdNickname;
 import project.flipnote.user.entity.UserProfile;
 import project.flipnote.user.entity.UserStatus;
